`RftPlotter` bug fixes and improvements (#854)
* refactored code into files for layout, callbacks and business logic

* fixed deselecting all ensembles bug

* Fixed problem with no data matching filters

* added title to the formations plot

* Removed one date mark from the slider in the map controls

* simplified title placement

* updated links in documentation

* fixed the problem of map plot changing size
