+++ CHANGELOG.md
@@ -9,13 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [#856](https://github.com/equinor/webviz-subsurface/pull/856) - `VolumetricAnalysis` - Added support for comparing sensitivities both within and across ensembles.
 - [#721](https://github.com/equinor/webviz-subsurface/pull/721) - Added data provider for reading ensemble summary data through a unified interface, supporting optional lazy resampling/interpolation depending on data input format.
-- [#845](https://github.com/equinor/webviz-subsurface/pull/845) - Added realization plot colored by sensitivity to tornado tab in `VolumetricAnalysis`. 
+- [#845](https://github.com/equinor/webviz-subsurface/pull/845) - Added realization plot colored by sensitivity to tornado tab in `VolumetricAnalysis`.
 
 ### Changed
 - [#855](https://github.com/equinor/webviz-subsurface/pull/855) - `VolumetricAnalysis` now supports mixing sensitivity and non-sensitivity ensembles.
 - [#853](https://github.com/equinor/webviz-subsurface/pull/853) - `ParameterResponseCorrelation` improvements. Constant parameters are removed from the correlation figure, and option to set maximum number of parameters is added. Trendline is added to the scatterplot. Axis in correlation figure is now calculated based on data.
 - [#844](https://github.com/equinor/webviz-subsurface/pull/844) - `SeismicMisfit` improvements. Data ranges now follows selected attribute. User defined zooms are now kept during callbacks. New option in slice plot to show individual realizations. Prettyfied all hoverdata. New colorscales. Polygons sorted by name in drop down selector.
 - [#842](https://github.com/equinor/webviz-subsurface/pull/842) - `GroupTree` improvements. Supporting groups as leaf nodes.
+- [#854](https://github.com/equinor/webviz-subsurface/pull/854) - `RFTPlotter` improvements. Fixed some bugs that caused webviz to crash, improved the layout some places and fixed broken links in documentation.
 
 ## [0.2.7] - 2021-11-08
