+++ webviz_subsurface/plugins/_rft_plotter/_errorplot_figure.py
@@ -1,14 +1,10 @@
-from typing import Dict, Union
+from typing import Dict
 
 import pandas as pd
+import webviz_core_components as wcc
 
 
-def update_errorplot(
-    df: pd.DataFrame, enscolors: Dict[str, str]
-) -> Dict[
-    str, Union[list, dict]
-]:  # TODO(Ruben Thoms) Better to make use of a TypedDict here at some point.
-
+def update_errorplot(df: pd.DataFrame, enscolors: Dict[str, str]) -> wcc.Graph:
     df["RFT_NAME"] = df.agg(
         lambda x: f"{x['WELL']} {int(x['YEAR'])} {x['ZONE']} ({int(x['TVD'])} TVD)",
         axis=1,
@@ -39,4 +35,6 @@ def update_errorplot(
         "xaxis": {"title": "Difference in bar"},
     }
 
-    return {"data": traces, "layout": layout}
+    return wcc.Graph(
+        style={"height": "84vh"}, figure={"data": traces, "layout": layout}
+    )
