+++ webviz_subsurface/plugins/_rft_plotter/_formation_figure.py
@@ -56,11 +56,11 @@ class FormationFigure:
                     (self.pressure_range[1] + self.xaxis_extension),
                 ],
             },
-            "height": 800,
             "legend": {"orientation": "h"},
             "margin": {"t": 50},
             "hovermode": "closest",
             "uirevision": f"{well}{depth_option}",
+            "title": well,
         }
 
     @property
