+++ webviz_subsurface/plugins/_rft_plotter/_map_figure.py
@@ -48,7 +48,6 @@ class MapFigure:
                     "cmax": self.ertdf[colorby].quantile(0.9),
                     "colorscale": [[0, "#2584DE"], [1, "#E50000"]],
                     "showscale": True,
-                    "colorbar": {"x": -0.1},
                 },
             }
         )
@@ -73,9 +72,8 @@ class MapFigure:
         return {
             "hovermode": "closest",
             "legend": {"itemsizing": "constant", "orientation": "h"},
-            "height": 800,
             "colorway": ["red", "blue"],
-            "margin": {"t": 50, "l": 0, "r": 0},
+            "margin": {"t": 50, "l": 50},
             "xaxis": {"constrain": "domain", "showgrid": False},
             "yaxis": {"scaleanchor": "x", "showgrid": False},
         }
